splitting js
