+++ src/index.js
@@ -3,7 +3,9 @@ require("@babel/register")({
   "plugins": [["import", {"libraryName": "@material-ui/core"}], "@babel/plugin-syntax-dynamic-import"]
 });
 import './js/HeathScript';
+// import './js/modules/theme-switcher';
+import './js/modules/nav-scroller';
 import './js/modules/show-more-fadebar/show-more.js';
+import './js/modules/hs-timeline';
 import './js/contact_me';
 import './js/jqBootstrapValidation';
-import './js/swiper';
