+++ src/js/modules/theme-switcher/index.js
@@ -1,0 +1,6 @@
+require("@babel/register")({
+  presets: ["@babel/preset-env"],
+  "plugins": [["import", {"libraryName": "@material-ui/core"}], "@babel/plugin-syntax-dynamic-import"]
+});
+
+import './theme-switcher';
